reorganizing the repo structure and splitting up the funcs.py script into separate themed scripts
